refactor: port destination-blurb to svelte
